Fix timeout errors smartly
