+++ x-pack/platform/plugins/shared/dataset_quality/server/services/data_telemetry/data_telemetry_service.ts
@@ -67,29 +67,6 @@ function isLicenseError(error: any): boolean {
   );
 }
 
-/**
- * Checks if an error is a timeout error (request took too long)
- */
-function isTimeoutError(error: any): boolean {
-  const errorType = error?.constructor?.name;
-  const errorName = error?.name;
-  const errorMessage = error?.message?.toLowerCase() || '';
-
-  return (
-    errorType === 'TimeoutError' ||
-    errorName === 'TimeoutError' ||
-    errorMessage.includes('timeout') ||
-    errorMessage.includes('timed out')
-  );
-}
-
-/**
- * Checks if an error is expected during telemetry collection and should be logged as debug
- */
-function isExpectedTelemetryError(error: any): boolean {
-  return isLicenseError(error) || isTimeoutError(error);
-}
-
 export class DataTelemetryService {
   private readonly logger: Logger;
 
@@ -202,12 +179,11 @@ export class DataTelemetryService {
               } catch (e) {
                 if (e.message === SKIP_COLLECTION) {
                   data = null; // Collection is skipped, skip reporting
-                } else if (isExpectedTelemetryError(e)) {
-                  // Expected errors (license restrictions, timeouts) are operational issues,
-                  // not system problems, so they should be logged as debug
-                  const errorType = isLicenseError(e) ? 'license restriction' : 'timeout';
+                } else if (isLicenseError(e)) {
+                  // License errors are expected in certain deployments (e.g., basic license)
+                  // and don't indicate a problem with the system
                   service.logger.debug(
-                    `[Logs Data Telemetry] Skipping telemetry collection due to ${errorType}: ${e.message}`
+                    `[Logs Data Telemetry] Skipping telemetry collection due to license restriction: ${e.message}`
                   );
                   data = null;
                 } else {
@@ -272,6 +248,7 @@ export class DataTelemetryService {
         logsIndexPatterns: LOGS_DATASET_INDEX_PATTERNS,
         excludeStreamsStartingWith: [...NON_LOG_SIGNALS, ...EXCLUDE_ELASTIC_LOGS],
         breatheDelay: BREATHE_DELAY_MEDIUM,
+        logger: this.logger,
       }).pipe(
         switchMap((dataStreamsAndIndicesInfo) => {
           if (dataStreamsAndIndicesInfo.length > MAX_STREAMS_TO_REPORT) {
